--- app/src/main/java/com/dilekbaykara/tasky/auth/util/Validation.kt
@@ -1,2 +1,0 @@
-package com.dilekbaykara.tasky.auth.util
-
