+++ app/src/main/java/com/dilekbaykara/tasky/domain/model/AuthModels.kt
@@ -1,4 +1,4 @@
-package com.dilekbaykara.tasky.auth.models
+package com.dilekbaykara.tasky.domain.model
 
 
 import com.google.gson.annotations.SerializedName
