+++ app/src/main/java/com/dilekbaykara/tasky/presentation/auth/error/SuccessDialog.kt
@@ -1,10 +1,9 @@
-package com.dilekbaykara.tasky.auth.dialogs
+package com.dilekbaykara.tasky.presentation.auth.error
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import com.dilekbaykara.tasky.auth.ui.Error
 
 @Composable
 fun SuccessDialog(title: String, message: String, onDismiss: () -> Unit) {
