+++ app/src/main/java/com/dilekbaykara/tasky/presentation/auth/register/AuthUX.kt
@@ -1,4 +1,4 @@
-package com.dilekbaykara.tasky.auth.ui
+package com.dilekbaykara.tasky.presentation.auth.register
 
 import android.widget.Toast
 import androidx.compose.foundation.clickable
@@ -30,8 +30,7 @@ import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.dilekbaykara.tasky.auth.dialogs.ErrorDialog
-import com.dilekbaykara.tasky.auth.dialogs.SuccessDialog
+import com.dilekbaykara.tasky.presentation.auth.error.SuccessDialog
 
 @Composable
 fun Header(modifier: Modifier = Modifier) {
