+++ app/src/main/java/com/dilekbaykara/tasky/presentation/common/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.dilekbaykara.tasky.ui.theme
+package com.dilekbaykara.tasky.presentation.common.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
