+++ app/src/main/java/com/dilekbaykara/tasky/presentation/splash/SplashViewModel.kt
@@ -1,4 +1,4 @@
-package com.dilekbaykara.tasky.splash
+package com.dilekbaykara.tasky.presentation.splash
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
